Criação da classe de conexao com o banco de dados e os metodos de CRUD, tambem aproveitei para refatorar as classes para uma melhor organização
