+++ src/Form/RelatorioEspacoForm.java
@@ -10,7 +10,7 @@ import java.awt.event.WindowEvent;
 import java.util.List;
 import java.util.Vector;
 
-import static View.ViewController.chamaMenuRelatorios;
+import static Controller.ViewController.chamaMenuRelatorios;
 
 public class RelatorioEspacoForm extends JPanel {
