+++ src/Form/RelatorioEsporteForm.java
@@ -11,7 +11,7 @@ import java.awt.event.WindowEvent;
 import java.util.List;
 import java.util.Vector;
 
-import static View.ViewController.chamaMenuRelatorios;
+import static Controller.ViewController.chamaMenuRelatorios;
 
 public class RelatorioEsporteForm extends JPanel {
