+++ src/Form/RelatorioMaterialForm.java
@@ -1,7 +1,6 @@
 package Form;
 
 import Model.Material;
-import Relatorio.TableCliente;
 import Relatorio.TableMaterial;
 
 import javax.swing.*;
@@ -12,7 +11,7 @@ import java.awt.event.WindowEvent;
 import java.util.List;
 import java.util.Vector;
 
-import static View.ViewController.chamaMenuRelatorios;
+import static Controller.ViewController.chamaMenuRelatorios;
 
 public class RelatorioMaterialForm extends JPanel {
