+++ src/Main.java
@@ -1,4 +1,4 @@
-import View.View;
+import Model.View;
 
 public class Main extends View {
     public static void main(String[] args) {
