+++ src/Model/Calcular.java
@@ -1,5 +1,4 @@
 package Model;
-
 import java.math.BigDecimal;
 import java.math.RoundingMode;
