+++ src/Model/CalcularCodigo.java
@@ -1,5 +1,4 @@
 package Model;
-
 import java.util.List;
 
 public class CalcularCodigo {
