+++ src/Repository/CidadeDAO.java
@@ -1,12 +1,10 @@
 package Repository;
-
 import Model.*;
-
 import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
-
 import static Model.VerificaRegistroNullo.verificaRegistroNullo;
+import static Model.View.chamaMenuPrincipal;
 
 public class CidadeDAO {
     static List<Cidade> cidades = new ArrayList<>();
@@ -59,5 +57,31 @@ public class CidadeDAO {
 
     public static List<Cidade> getCidades() {
         return cidades;
+    }
+    public static void cadastroCidade() {
+        try {
+
+            if (EstadoDAO.getEstados().size() == 0) {
+                JOptionPane.showMessageDialog(null, "Nenhum estado cadastrado!" + "\n" + "Por favor cadastre um estado.", "Atenção", JOptionPane.WARNING_MESSAGE);
+                chamaMenuPrincipal();
+            }
+
+            Integer codigoCidade = CalcularCodigo.calculaCodigo(CidadeDAO.getCidades());
+            String nomeCidade = JOptionPane.showInputDialog(null, "Digite o nome da cidade");
+            verificaRegistroNullo(nomeCidade);
+            Object[] selectionValuesEstado = EstadoDAO.findEstadoInArray();
+            String initialSelectionEstado = (String) selectionValuesEstado[0];
+            Object selectionEstado = JOptionPane.showInputDialog(null, "Selecione o estado",
+                    "Lista de Estados", JOptionPane.QUESTION_MESSAGE, null, selectionValuesEstado, initialSelectionEstado);
+            List<Estado> estados = EstadoDAO.buscarPorNome(selectionEstado.toString());
+            verificaRegistroNullo(selectionEstado);
+            Cidade cidade = new Cidade(codigoCidade, nomeCidade, estados.get(0));
+            CidadeDAO.salvar(cidade);
+            JOptionPane.showMessageDialog(null, "Cadastro salvo com sucesso!", "Sucesso", JOptionPane.INFORMATION_MESSAGE);
+            chamaMenuPrincipal();
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(null, "Cadastro Invalido, favor tentar novamente!", "Erro", JOptionPane.ERROR_MESSAGE);
+            chamaMenuPrincipal();
+        }
     }
 }
