+++ src/Repository/ClienteDAO.java
@@ -1,11 +1,13 @@
 package Repository;
+import Model.CalcularCodigo;
 import Model.Cliente;
-
 import javax.swing.*;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
-
 import static Model.VerificaRegistroNullo.verificaRegistroNullo;
+import static Model.View.chamaMenuPrincipal;
 
 public class ClienteDAO {
     static List<Cliente> clientes = new ArrayList<>();
@@ -61,5 +63,34 @@ public class ClienteDAO {
 
     public static List<Cliente> getClientes() {
         return clientes;
+    }
+
+    public static void cadastroCliente() {
+        try {
+            Integer codigo = CalcularCodigo.calculaCodigo(ClienteDAO.getClientes());
+            String nome = JOptionPane.showInputDialog(null, "Digite o nome");
+            verificaRegistroNullo(nome);
+            LocalDate nascimento = null;
+            String imputnascimento = JOptionPane.showInputDialog(null, "Digite a data de nascimento");
+            try {
+                nascimento = LocalDate.parse(imputnascimento, DateTimeFormatter.ofPattern("dd/MM/yyyy"));
+            } catch (Exception e) {
+                JOptionPane.showMessageDialog(null, "Data invalida, tente no formato dd/MM/yyyy");
+            }
+            verificaRegistroNullo(nascimento);
+            String telefone = JOptionPane.showInputDialog(null, "Digite o telefone");
+            verificaRegistroNullo(telefone);
+            String cpf = JOptionPane.showInputDialog(null, "Digite o cpf");
+            verificaRegistroNullo(cpf);
+            String rg = JOptionPane.showInputDialog(null, "Digite o rg");
+            verificaRegistroNullo(rg);
+            Cliente pessoa = new Cliente(codigo, nome, nascimento, telefone, cpf, rg);
+            ClienteDAO.salvar(pessoa);
+            JOptionPane.showMessageDialog(null, "Cadastro salvo com sucesso!", "Sucesso", JOptionPane.INFORMATION_MESSAGE);
+            chamaMenuPrincipal();
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(null, "Cadastro Invalido, favor tentar novamente!", "Erro", JOptionPane.ERROR_MESSAGE);
+            chamaMenuPrincipal();
+        }
     }
 }
